+++ src/locale/hr/index.js
@@ -4,6 +4,7 @@ var buildFormatLocale = require('./build_format_locale/index.js')
 /**
  * @category Locales
  * @summary Croatian locale.
+ * @author Matija Marohnić [@silvenon]{@link https://github.com/silvenon}
  */
 module.exports = {
   distanceInWords: buildDistanceInWordsLocale(),