+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This change log follows the format documented in [Keep a CHANGELOG].
 
 ## Added
 
+- Auto generate typescript definitions from documentation on release
+
 - [Croatian locale (hr)](https://github.com/date-fns/date-fns/pull/362)
   (thanks to Matija Marohnić [@silvenon](https://github.com/silvenon))